### 2022-08-11 #4667 @editorialbot
[PRE REVIEW]: tpcp: Tiny Pipelines for Complex Problems - A set of framework independent helpers for algorithms development and evaluation
**Submitting author:** @AKuederle (<a href="http://orcid.org/0000-0002-5686-281X">Arne Küderle</a>)
**Repository:** https://github.com/mad-lab-fau/tpcp/
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.0
