### 2022-08-11 #4667 @danielskatz
@AKuederle -  Please fix the invalid DOI that editorialbot mentions. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, in the paper, when you reference a citation directly, you should remove the []s in the md. For example, `published by Roth et al. [@Roth2021a]` should be `published by @Roth2021a`  And finally, you seem to be missing a References heading before your references.
