+++ paper/tpcp.bib
@@ -58,22 +58,23 @@
   publisher = {{Curran Associates, Inc.}}
 }
 
-@article{Pedregosa2011a,
+@misc{Pedregosa2018,
   title = {Scikit-Learn: {{Machine Learning}} in {{Python}}},
-  author = {Pedregosa, Fabian and Varoquaux, Gael and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, M. and Duchesnay, E.},
-  year = {2011},
-  journal = {Journal of Machine Learning Research},
-  volume = {12},
-  eprint = {1201.0490v2},
+  shorttitle = {Scikit-Learn},
+  author = {Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and M{\"u}ller, Andreas and Nothman, Joel and Louppe, Gilles and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and Vanderplas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'E}douard},
+  year = {2018},
+  month = jun,
+  number = {arXiv:1201.0490},
+  eprint = {1201.0490},
   eprinttype = {arxiv},
-  pages = {2825--2830},
-  issn = {15324435},
-  doi = {10.1007/s13398-014-0173-7.2},
-  abstract = {Scikit-learn is a Python module integrating a wide range of state-of-the-art machine learning algo-rithms for medium-scale supervised and unsupervised problems. This package focuses on bring-ing machine learning to non-specialists using a general-purpose high-level language. Emphasis is put on ease of use, performance, documentation, and API consistency. It has minimal dependen-cies and is distributed under the simplified BSD license, encouraging its use in both academic and commercial settings. Source code, binaries, and documentation can be downloaded from http://scikit-learn.sourceforge.net.},
+  primaryclass = {cs},
+  publisher = {{arXiv}},
+  doi = {10.48550/arXiv.1201.0490},
+  url = {http://arxiv.org/abs/1201.0490},
+  urldate = {2022-08-11},
+  abstract = {Scikit-learn is a Python module integrating a wide range of state-of-the-art machine learning algorithms for medium-scale supervised and unsupervised problems. This package focuses on bringing machine learning to non-specialists using a general-purpose high-level language. Emphasis is put on ease of use, performance, documentation, and API consistency. It has minimal dependencies and is distributed under the simplified BSD license, encouraging its use in both academic and commercial settings. Source code, binaries, and documentation can be downloaded from http://scikit-learn.org.},
   archiveprefix = {arXiv},
-  isbn = {9781783281930},
-  pmid = {1000044560},
-  keywords = {model selection,Python,supervised learning,unsupervised learning}
+  keywords = {Computer Science - Machine Learning,Computer Science - Mathematical Software}
 }
 
 @article{Roth2021a,
