### 2022-09-08 e0c3838
Changed pyright python verison
