### 2022-09-08 df04b5a
Fixed issue that empty classes would not pass the tpcp checks
