### 2022-09-09 0e724bb
Fixed tests affected by change in annotation processing order
