+++ tests/test_parameter_string_annot.py
@@ -26,7 +26,7 @@ def test_basic_annotation_collection():
             self.custom_annotated = custom_annotated
             self.normal_no_annot = normal_no_annot
 
-    # We manually trigger the processing of the annotations
+    # We manually trigger the second part of the processing of the annotations
     _process_tpcp_class(Test)
 
     assert Test.__field_annotations__ == {
@@ -55,7 +55,7 @@ def test_import_forward():
             self.custom_annotated = custom_annotated
             self.normal_no_annot = normal_no_annot
 
-    # We manually trigger the processing of the annotations
+    # We manually trigger the second part of the processing of the annotations
     _process_tpcp_class(Test)
 
     assert Test.__field_annotations__ == {
@@ -70,22 +70,18 @@ def test_import_forward_error():
     if TYPE_CHECKING:
         from tpcp import optimize
 
-    class Test(BaseTpcpObject):
-        hyper: HyperPara[int]
-        normal: Para[optimize.GridSearch]
-        custom_annotated: Annotated[HyperPara[int], "custom_metadata"]
-        normal_no_annot: int
-
-        def __init__(self, hyper: int, normal: optimize.GridSearch, custom_annotated: int, normal_no_annot: int):
-            self.hyper = hyper
-            self.normal = normal
-            self.custom_annotated = custom_annotated
-            self.normal_no_annot = normal_no_annot
-
     with pytest.raises(RuntimeError) as e:
+        class Test(BaseTpcpObject):
+            hyper: HyperPara[int]
+            normal: Para[optimize.GridSearch]
+            custom_annotated: Annotated[HyperPara[int], "custom_metadata"]
+            normal_no_annot: int
 
-        # We manually trigger the processing of the annotations
-        _process_tpcp_class(Test)
+            def __init__(self, hyper: int, normal: optimize.GridSearch, custom_annotated: int, normal_no_annot: int):
+                self.hyper = hyper
+                self.normal = normal
+                self.custom_annotated = custom_annotated
+                self.normal_no_annot = normal_no_annot
 
     assert "You ran into an edge case" in str(e)
 
