### 2022-09-09 6c7aee6
Reworked how validation and annotation extractions works once again
