+++ tpcp/_base.py
@@ -161,12 +161,17 @@ def _retry_eval_with_missing_locals(
 
 
 def _custom_get_type_hints(cls: Type[_BaseTpcpObject]) -> Dict[str, Any]:
-    """Extract type hints while avoiding issues with forward references."""
+    """Extract type hints while avoiding issues with forward references.
+
+    We automatically skip all douple_underscore methods.
+    """
     hints = {}
     for base in reversed(cls.__mro__):
         base_globals = sys.modules[base.__module__].__dict__
         ann = base.__dict__.get("__annotations__", {})
         for name, value in ann.items():
+            if name.startswith("__"):
+                continue
             if value is None:
                 value = type(None)
             elif isinstance(value, str):
@@ -177,7 +182,9 @@ def _custom_get_type_hints(cls: Type[_BaseTpcpObject]) -> Dict[str, Any]:
     return hints
 
 
-def _extract_partial_annotations(cls: Type[_BaseTpcpObject]) -> Dict[str, _ParaTypes]:
+def _extract_annotations(
+    cls: Type[_BaseTpcpObject], init_fields: Dict[str, inspect.Parameter]
+) -> Dict[str, _ParaTypes]:
     cls_annotations = _custom_get_type_hints(cls)
     para_annotations = {}
     for k, v in cls_annotations.items():
@@ -191,36 +198,21 @@ def _extract_partial_annotations(cls: Type[_BaseTpcpObject]) -> Dict[str, _ParaT
                 if isinstance(annot, _ParaTypes):
                     para_annotations[k] = annot
                     break
-
+        elif k in init_fields:
+            para_annotations[k] = _ParaTypes.SIMPLE
     return para_annotations
 
 
-def _add_missing_init_fields_to_annotations(
-    para_annotations: Dict[str, _ParaTypes], init_fields: Dict[str, inspect.Parameter]
-):
-    for k in init_fields:
-        if k not in para_annotations:
-            para_annotations[k] = _ParaTypes.SIMPLE
-
-
-def _run_field_checks(cls, fields: Dict[str, Any]) -> None:
-    _has_dangerous_mutable_default(fields, cls)
-    _has_invalid_name(fields, cls)
-    _annotations_are_valid(fields, cls)
-
-
-def _process_tpcp_class(cls: Type[_BaseTpcpObject]):
+def _validate_parameter(instance: _BaseTpcpObject):
     # We extract the fields of the init
-    fields = _get_init_defaults(cls)
-
-    # Now that we know the init fields for sure, we can complete our parameter extraction
-    _add_missing_init_fields_to_annotations(cls.__field_annotations__, fields)
+    fields = _get_init_defaults(type(instance))
 
     # Validation
-    if cls.__skip_validation__ is not True:
-        _run_field_checks(cls, fields)
+    if instance.__skip_validation__ is not True:
+        _has_dangerous_mutable_default(fields, instance)
+        _has_invalid_name(fields, instance)
 
-    if dataclasses.is_dataclass(cls) and any(isinstance(field.default, BaseFactory) for field in fields.values()):
+    if dataclasses.is_dataclass(instance) and any(isinstance(field.default, BaseFactory) for field in fields.values()):
         # If the class is already a data class when this check runs (aka when it is triggered by the custom
         # decorator), then we don't need to do anything.
         # The BaseTpcpObject already implements the post_init method to handly all factory stuff.
@@ -231,42 +223,36 @@ def _process_tpcp_class(cls: Type[_BaseTpcpObject]):
             "Consider using the `factory` option of `dataclasses.field` instead."
         )
 
-    cls.__tpcp_cls_processed__ = True
-
-
[…]
