### 2022-09-09 006cf72
Added additional special case handling when inits are already wrapped
