+++ .readthedocs.yml
@@ -14,12 +14,7 @@ build:
   tools:
     python: "3.8"
   jobs:
-    pre_create_environment:
-      - asdf plugin add poetry
-      - asdf install poetry latest
-      - asdf global poetry latest
-      - poetry config virtualenvs.create false
     post_install:
-      - poetry config cache-dir .
-      - poetry env use python
+      - pip install poetry
+      - poetry config virtualenvs.create false
       - poetry install -E torch -E optuna
