### 2022-09-12 57b96b8
Rudimentary first implementation of an Info object concept
