### 2022-09-15 5382e9a
CustomOptunaOptimize now expects a callable to create the study not a study object itself
