+++ pyproject.toml
@@ -21,7 +21,7 @@ pandas = "^1"
 tqdm = "^4.62.3"
 typing-extensions = ">=4.1.1"
 torch = {version = ">=1.6.0", optional = true}
-optuna = {version = "^2.10.0", optional = true}
+optuna = {version = "^2.10.0 | ^3.0.2", optional = true}
 
 [tool.poetry.extras]
 torch = ["torch"]
