### 2022-09-19 a229e7f
dataclass is normal class by default again
