+++ tpcp/_dataset.py
@@ -11,7 +11,6 @@ from tpcp._base import BaseTpcpObject
 DatasetT = TypeVar("DatasetT", bound="Dataset")
 
 
-@dataclasses.dataclass
 class Dataset(BaseTpcpObject, _skip_validation=True):
     """Baseclass for tpcp Dataset objects.
 
@@ -191,8 +190,17 @@ class Dataset(BaseTpcpObject, _skip_validation=True):
 
     """
 
-    groupby_cols: Optional[Union[List[str], str]] = None
-    subset_index: Optional[pd.DataFrame] = None
+    groupby_cols: Optional[Union[List[str], str]]
+    subset_index: Optional[pd.DataFrame]
+
+    def __init__(  # noqa: super-init-not-called
+        self,
+        *,
+        groupby_cols: Optional[Union[List[str], str]] = None,
+        subset_index: Optional[pd.DataFrame] = None,
+    ):
+        self.groupby_cols = groupby_cols
+        self.subset_index = subset_index
 
     @property
     def index(self) -> pd.DataFrame:
