### 2022-09-19 174df9d
PRovide conversion function for the use of dataclasses or attrs for important classes
