+++ tpcp/_dataset.py
@@ -1,6 +1,6 @@
 """Base class for all datasets."""
 import dataclasses
-from typing import Dict, Iterator, List, Optional, Sequence, Tuple, TypeVar, Union, cast, overload
+from typing import Dict, Iterator, List, Optional, Sequence, Tuple, TypeVar, Union, cast, overload, Type
 
 import numpy as np
 import pandas as pd
@@ -8,10 +8,351 @@ from typing_extensions import Self
 
 from tpcp._base import BaseTpcpObject
 
-DatasetT = TypeVar("DatasetT", bound="Dataset")
+DatasetT = TypeVar("DatasetT", bound="_Dataset")
 
 
-class Dataset(BaseTpcpObject, _skip_validation=True):
+class _Dataset(BaseTpcpObject, _skip_validation=True):
+    groupby_cols: Optional[Union[List[str], str]]
+    subset_index: Optional[pd.DataFrame]
+
+    @property
+    def index(self) -> pd.DataFrame:
+        """Get index."""
+        if self.subset_index is None:
+            return self.create_index()
+
+        return self.subset_index
+
+    @property
+    def groups(self) -> List[Union[str, Tuple[str, ...]]]:
+        """Get all groups based on the set groupby level.
+
+        This will either return a list of strings/integers, if there is only a single group level or index column.
+        If there are multiple groupy levels/index columns, it will return a list of named tuples.
+
+        Note, that if one of the groupby levels/index columns is not a valid Python attribute name (e.g. in contains
+        spaces or starts with a number), the named tuple will not contain the correct column name!
+        For more information see the documentation of the `rename` parameter of :func:`collections.namedtuple`.
+        """
+        if len(self._get_groupby_columns()) == 1:
+            return self._get_unique_groups().to_list()
+        return list(self._get_unique_groups().to_frame().itertuples(index=False, name=type(self).__name__))
+
+    @property
+    def group(self) -> Union[str, Tuple[str, ...]]:
+        """Get the current group.
+
+        Note, this attribute can only be used, if there is just a single group.
+        If there is only a single groupby column or column in the index, this will return a string.
+        Otherwise, this will return a named tuple.
+        """
+        self.assert_is_single_group("group")
+        return self.groups[0]
+
+    def __len__(self) -> int:
+        """Get the length of the dataset.
+
+        This is equal to the number of rows in the index, if `self.groupby_cols=None`.
+        Otherwise, it is equal to the number of unique groups.
+        """
+        return len(self.groups)
+
+    @property
+    def shape(self) -> Tuple[int]:
+        """Get the shape of the dataset.
+
+        This only reports a single dimension.
+        This is equal to the number of rows in the index, if `self.groupby_cols=None`.
+        Otherwise, it is equal to the number of unique groups.
+        """
+        return (len(self),)
+
+    @property
+    def grouped_index(self) -> pd.DataFrame:
+        """Return the index with the `groupby` columns set as multiindex."""
+        groupby_cols = self._get_groupby_columns()
+        try:
+            return self.index.set_index(groupby_cols, drop=False)
+        except KeyError as e:
+            raise KeyError(
+                f"You can only groupby columns that are part of the index columns ({list(self.index.columns)}) and not"
+                f" {self.groupby_cols}"
+            ) from e
+
+    def _get_groupby_columns(self) -> List[str]:
+        """Get the groupby columns."""
+        if self.groupby_cols is None:
+            return self.index.columns.to_list()
+        return _ensure_is_list(self.groupby_cols)
+
+    def _get_unique_groups(self) -> Union[pd.MultiIndex, pd.Index]:
+        return self.grouped_index.index.unique()
+
[…]
