+++ examples/other_features/_01_dataclasses.py
@@ -50,6 +50,9 @@ from dataclasses import dataclass, field
 # The only downside is that the annotation of result fields and constants is a little more verbose, and you **need** to
 # make sure that these parameters are excluded from the init.
 # Otherwise, tpcp will explode ;)
+#
+# Note, if you are using Python >=3.10, we highly recommend to use the `kw_only` option for dataclasses,
+# which prevent some of the inheritance issues of dataclasses.
 from typing import ClassVar
 
 
@@ -82,11 +85,10 @@ QRSDetector(high_pass_filter_cutoff_hz=4, max_heart_rate_bpm=200, min_r_peak_hei
 # Instead of repeating all parameters, you just need to specify the new once.
 # However, you need to make sure that you also apply the `dataclass` decorator to the child class!
 #
-# But, be aware that you can not subclass a class that is not a `dataclass` and also has a `__init__` method!
-# For example, you can not subclass :class:`~tpcp.optimize.GridSearch` with a dataclass, as it already defines its own
-# `__init__`.
-# In this case you need to use a regular class and manually repeat all parent parameters
-# (and call `super().__init__()`).
+# ... warning :: New parameters will be added at the end in the positional order in the init method.
+#                To avoid passing the wrong values to the wrong parameters, we highly recommend to pass parameters
+#                only by name and not by position, or use the `kw_only` parameter of dataclasses supported in Python
+#                >=3.10.
 @dataclass
 class ModifiedQRSDetector(QRSDetector):
     new_parameter: Parameter[float] = 3
@@ -96,6 +98,37 @@ ModifiedQRSDetector(
     high_pass_filter_cutoff_hz=4, max_heart_rate_bpm=200, min_r_peak_height_over_baseline=1, new_parameter=3
 )
 
+# %%
+# Inheritance from complex tpcp classes
+# --------------------------------------
+# While inheriting from other dataclasses works without issues, be aware that you can not subclass a class that is
+# not a `dataclass` and also has a `__init__` method!
+# For example, you can not subclass :class:`~tpcp.optimize.GridSearch` with a dataclass, as it already defines its own
+# `__init__`.
+# In this case you need to use a regular class and manually repeat all parent parameters
+# (and call `super().__init__()`).
+#
+# While this might not be a big deal for the GridSearch class, as you are not expected to subclass it on a regular, it
+# can become annoying for classes like `~tpcp.Dataset` and `~tpcp.optimize.optuna.CustomOptunaOptimize`,
+# which already have an init and you need to subclass to work with them.
+# For these two classes (and other classes with predefined inits, we expect you to subclass from), we provide a
+# `as_dataclass` class method that returns a data class version of the respective class:
+from tpcp import Dataset
+from itertools import product
+
+
+@dataclass()
+class CustomDataset(Dataset.as_dataclass()):  # Note the `as_dataclass` call here!
+    def create_index(self) -> pd.DataFrame:
+        return pd.DataFrame(
+            list(product(("patient_1", "patient_2", "patient_3"), ("test_1", "test_2"), ("1", "2"))),
+            columns=["patient", "test", "extra"],
+        )
+
+    custom_param: float = 2  # This must have a default value, as the baseclass has parameters with defautls
+
+
+CustomDataset(custom_param=3)
 
 # %%
 # Mutable Defaults
@@ -107,7 +140,7 @@ ModifiedQRSDetector(
 
 @dataclass
 class FilterAlgorithm(Algorithm):
[…]
