+++ pyproject.toml
@@ -22,10 +22,12 @@ tqdm = "^4.62.3"
 typing-extensions = ">=4.1.1"
 torch = {version = ">=1.6.0", optional = true}
 optuna = {version = "^2.10.0 | ^3.0.2", optional = true}
+attrs = {version = "^22.1.0", optional = true}
 
 [tool.poetry.extras]
 torch = ["torch"]
 optuna = ["optuna"]
+attrs = ["attrs"]
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.1"

[…]
