### 2022-09-19 cb12691
Reworked caching slightly to now rely on __init__subclass__
