### 2022-09-19 ff56d8d
Optuna class is now a generic once again
