### 2022-09-19 5bd05f1
Properly handled validation and annotations without relying on __init_subclass__
