+++ tpcp/_base.py
@@ -97,8 +97,15 @@ def _replace_defaults_wrapper(old_init: Callable) -> Callable:
     def new_init(self: BaseTpcpObject, *args: Any, **kwargs: Any) -> None:
         # call the old init.
         old_init(self, *args, **kwargs)
-        self.__post_init__()
 
+        # Check if any of the initial values has a "default parameter flag".
+        # If yes we replace it with a clone (in case of a tpcp object) or a deepcopy in case of other objects.
+        # This is handled by the factory `get_value` method.
+        for p in get_param_names(self):
+            if isinstance(val := getattr(self, p), BaseFactory):
+                setattr(self, p, val.get_value())
+
+    # This is just for introspection, in case we want to know if we have a modified init.
     new_init.__tpcp_wrapped__ = True
 
     return new_init
@@ -210,26 +217,39 @@ def _validate_parameter(instance: _BaseTpcpObject):
     fields = _get_init_defaults(type(instance))
 
     # Validation
-    if instance.__skip_validation__ is not True:
-        _has_dangerous_mutable_default(fields, instance)
-        _has_invalid_name(fields, instance)
+    _has_dangerous_mutable_default(fields, instance)
+    _has_invalid_name(fields, instance)
 
-    if dataclasses.is_dataclass(instance) and any(isinstance(field.default, BaseFactory) for field in fields.values()):
-        # If the class is already a data class when this check runs (aka when it is triggered by the custom
-        # decorator), then we don't need to do anything.
-        # The BaseTpcpObject already implements the post_init method to handly all factory stuff.
-        # However, when you are already using dataclasses, you should probably use their factory methods.
-        warnings.warn(
+    if (
+        dataclasses.is_dataclass(instance)
+        or getattr(
+            instance, "__attrs_attrs__", False
+        )  # This checks if the class it an `attrs` class, without importing `attrs`
+    ) and any(isinstance(field.default, BaseFactory) for field in fields.values()):
+        # When you are already using dataclasses or attrs, you must use their factory methods, as we don't overwrite
+        # the init to support CloneFactory anymore.
+        raise ValidationError(
             "You are using the tpcp default factory (`cf`/`CloneFactory`) in combination with "
-            "`dataclasses`. "
-            "Consider using the `factory` option of `dataclasses.field` instead."
+            "`dataclasses` or `attrs`. "
+            "Use the `default_factory`/`factory` option of `dataclasses.field`/`attrs.field` instead."
         )
 
 
+def _get_tpcp_validated(cls_or_instance: Union[Type[_BaseTpcpObject], _BaseTpcpObject]):
+    cls = cls_or_instance if isinstance(cls_or_instance, type) else type(cls_or_instance)
+    return cls.__dict__.get("__tpcp_validated_hidden__", False)
+
+
+def _set_tpcp_validated(cls_or_instance: Union[Type[_BaseTpcpObject], _BaseTpcpObject], value: bool):
+    cls = cls_or_instance if isinstance(cls_or_instance, type) else type(cls_or_instance)
+    cls.__tpcp_validated_hidden__ = value
+
+
 class _BaseTpcpObject:
-    __field_annotations_cache__: Tuple[int, Dict[str, _ParaTypes]]
-    __skip_validation__: bool
-    __tpcp_cls_processed__: bool
+    # These two parameters should be initialized once per class.
+    # This means, when we read them, we always check in the dictionary if they are defined.
+    __field_annotations_cache__: Dict[str, _ParaTypes]
+    __tpcp_validated_hidden__: bool
 
     @property
     def __field_annotations__(self) -> Dict[str, _ParaTypes]:
@@ -241,10 +261,8 @@ class _BaseTpcpObject:
         cls = type(self)
[…]
