+++ tpcp/_dataset.py
@@ -150,7 +150,10 @@ class _Dataset(BaseTpcpObject, _skip_validation=True):
         """
         if (
             list(
-                map(lambda x: x is None or (isinstance(x, dict) and len(x) == 0), (groups, index, bool_map, kwargs),)
+                map(
+                    lambda x: x is None or (isinstance(x, dict) and len(x) == 0),
+                    (groups, index, bool_map, kwargs),
+                )
             ).count(False)
             > 1
         ):
@@ -532,7 +535,10 @@ class Dataset(_Dataset, _skip_validation=True):
     """
 
     def __init__(
-        self, *, groupby_cols: Optional[Union[List[str], str]] = None, subset_index: Optional[pd.DataFrame] = None,
+        self,
+        *,
+        groupby_cols: Optional[Union[List[str], str]] = None,
+        subset_index: Optional[pd.DataFrame] = None,
     ):
         self.groupby_cols = groupby_cols
         self.subset_index = subset_index
