+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "tpcp"
-version = "0.11.0a0"
+version = "0.11.0a1"
 description = "Pipeline and Dataset helpers for complex algorithm evaluation."
 authors = [
     "Arne Küderle <arne.kuederle@fau.de>",
