### 2022-09-19 1f1b95f
Removed type annot that prevented autocompletion from dataclass version
