### 2022-09-19 7ae6b93
Disabled slots for attrs classes
