### 2022-09-19 ee2b1ed
Removed wrong slot attribute from dataclass decorator
