+++ tpcp/optimize/optuna.py
@@ -348,7 +348,7 @@ class CustomOptunaOptimize(_CustomOptunaOptimize[PipelineT, DatasetT]):
     def as_dataclass():
         """Return a version of the Dataset class that can be subclassed using dataclasses."""
 
-        @dataclasses.dataclass(eq=False, repr=False, order=False, slots=False)
+        @dataclasses.dataclass(eq=False, repr=False, order=False)
         class CustomOptunaOptimizeDc(_CustomOptunaOptimize[PipelineT, DatasetT]):
             """Dataclass version of CustomOptunaOptimize."""
 
