### 2022-09-27 #4667 @AKuederle
brunaw, itismeghasyam, hayesall, gagolews, billchenxi, andreysmelter

@adi3 Unfortunately, it seems like no further people from the original list responded. Here are some more suggestions.

As a sidenote, the paper is not really related to robotics but more about fundamental ML and software engineering.

Let me know, if I can do anything further to move this further.
