### 2022-10-07 1690c2f
Implemented get_params to be aware of composite keys
