### 2022-10-07 9673bfc
Implemented get_params to be aware of composite keys
