### 2022-10-07 ae3484a
First implementation for set_params that support compound fields
