+++ tests/test_composite_objects.py
@@ -44,7 +44,9 @@ def test_composite_deep_get_params():
 )
 def test_raises_with_invalid_composite(values):
     test_instance = create_test_class(
-        "test", params={"nested": None}, private_params={"_composite_params": ("nested",)},
+        "test",
+        params={"nested": None},
+        private_params={"_composite_params": ("nested",)},
     )
     test_instance.set_params(nested=values[0])
     if values[1]:
@@ -56,7 +58,9 @@ def test_raises_with_invalid_composite(values):
 
 def test_set_params_simple():
     test_instance = create_test_class(
-        "test", params={"nested": None}, private_params={"_composite_params": ("nested",)},
+        "test",
+        params={"nested": None},
+        private_params={"_composite_params": ("nested",)},
     )
 
     test_instance.nested = [("bla", 2)]
@@ -68,7 +72,9 @@ def test_set_params_simple():
 
 def test_set_params_nested_obj():
     test_instance = create_test_class(
-        "test", params={"nested": None}, private_params={"_composite_params": ("nested",)},
+        "test",
+        params={"nested": None},
+        private_params={"_composite_params": ("nested",)},
     )
     nested_instance = create_test_class("nested", params={"nested1": "n1", "nested2": "n2"})
 
@@ -82,7 +88,9 @@ def test_set_params_nested_obj():
 def test_set_params_order():
     """This tests that we always set the least nested values first."""
     test_instance = create_test_class(
-        "test", params={"nested": None}, private_params={"_composite_params": ("nested",)},
+        "test",
+        params={"nested": None},
+        private_params={"_composite_params": ("nested",)},
     )
     nested_instance = create_test_class("nested", params={"nested1": "n1", "nested2": "n2"})
 
@@ -99,7 +107,9 @@ def test_set_params_order():
 
 def test_raises_when_setting_unknown():
     test_instance = create_test_class(
-        "test", params={"nested": None}, private_params={"_composite_params": ("nested",)},
+        "test",
+        params={"nested": None},
+        private_params={"_composite_params": ("nested",)},
     )
 
     test_instance.nested = [("bla", 2)]
