### 2022-10-17 6a426e9
Removed upper bound version dependencies to reduce chance of version conflicts
