+++ pyproject.toml
@@ -13,16 +13,16 @@ homepage = "https://github.com/mad-lab-fau/tpcp"
 repository = "https://github.com/mad-lab-fau/tpcp"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.11"
-joblib = "^1.1.0"
-scikit-learn = "^1"
-numpy = "^1"
-pandas = "^1"
-tqdm = "^4.62.3"
+python = ">=3.8,<4.0"
+joblib = ">=1.1"
+scikit-learn = ">=1.0"
+numpy = ">=1.0"
+pandas = ">=1.4"
+tqdm = ">=4.62.3"
 typing-extensions = ">=4.1.1"
 torch = {version = ">=1.6.0", optional = true}
-optuna = {version = "^2.10.0 | ^3.0.2", optional = true}
-attrs = {version = "^22.1.0", optional = true}
+optuna = {version = ">=2.10", optional = true}
+attrs = {version = ">=22.1.0", optional = true}
 
 [tool.poetry.extras]
 torch = ["torch"]
