### 2022-10-19 62fb792
Created first implementation of the `with_info` concept
