### 2022-10-19 c3c77aa
Created first implementation of the `with_info` concept
