### 2022-10-19 7d7e362
Added example for info object and fixed some bugs for implementation
