+++ examples/other_features/_03_optimization_info.py
@@ -1,0 +1,202 @@
+r"""
+.. _opti_info:
+
+Optimization Info
+=================
+
+Tpcp is focused on "running" pipelines and less on the "optimization" step.
+This is great for traditional algorithms and algorithms will complex return values, as you can easily store multiple
+parameters as attributes on the object during `run`.
+
+However, for optimization, you are limited to modifying input parameters.
+This works well in many cases, but sometimes, you need additional information from the optimization.
+For example, you might want to extract the loss decay of an iterative learning algorithms.
+This information is something that you wouldn't want to store in the input parameters (usually).
+
+For these cases tpcp provides the `self_optimize_with_info` method.
+This is basically identical to `self_optimize`, but is expected to provide two return values: the optimized instance
+AND an arbitrary additional object containing any information you like.
+Methods that get optimizable pipelines as input (e.g. :class:`~tpcp.optimize.Optimize` are aware of these method and
+will call `self_optimize_with_info` if available and store the additional info as result objects.
+
+The :class:`~tpcp.OptimizablePipeline` base-class is implemented in a way that you only need to worry about
+implementing either the `self_optimize_with_info` or the `self_optimize` method.
+The other will be available automatically (the additional info will be `NOTHING`, if the method is not implemented).
+
+If you are implementing a new Algorithm (instead of a pipeline), we don't provide this additional support,
+but it is relatively simple to implement yourself.
+
+In the following we will show how all of this works by expanding the QRS detection algorithm implemented in the other
+examples to return additional information from the optimization.
+"""
+from typing import List, Tuple, Dict
+
+import numpy as np
+import pandas as pd
+from sklearn.metrics import roc_curve
+from typing_extensions import Self
+
+from examples.algorithms.algorithms_qrs_detection_final import QRSDetector, match_events_with_reference, \
+    OptimizableQrsDetector
+from examples.datasets.datasets_final_ecg import ECGExampleData
+from tpcp import OptimizableParameter, HyperParameter, make_optimize_safe, OptimizablePipeline, Parameter, cf
+from tpcp.optimize import Optimize
+
+
+# %%
+# In the algorithm class below, we basically reimplemented the `OptimizableQrsDetector` from the algorithm example.
+# However, instead of the `self_optimize` method, we implemented the `self_optimize_with_info` method and added
+# additional information from the threshold selection process to the output of the optimization.
+#
+# To ensure interface compatibility with other algorithms, we also provided a `self_optimize` method, that simply calls
+# `self_optimize_with_info` under the hood.
+
+class OptimizableQrsDetectorWithInfo(QRSDetector):
+    min_r_peak_height_over_baseline: OptimizableParameter[float]
+    r_peak_match_tolerance_s: HyperParameter[float]
+
+    def __init__(
+        self,
+        max_heart_rate_bpm: float = 200.0,
+        min_r_peak_height_over_baseline: float = 1.0,
+        r_peak_match_tolerance_s: float = 0.01,
+        high_pass_filter_cutoff_hz: float = 1,
+    ):
+        self.r_peak_match_tolerance_s = r_peak_match_tolerance_s
+        super().__init__(
+            max_heart_rate_bpm=max_heart_rate_bpm,
+            min_r_peak_height_over_baseline=min_r_peak_height_over_baseline,
+            high_pass_filter_cutoff_hz=high_pass_filter_cutoff_hz,
+        )
+
+    @make_optimize_safe
+    def self_optimize_with_info(
[…]
