### 2022-10-19 90c8a32
Added example for info object and fixed some bugs for implementation
