### 2022-10-19 ea97320
Formatting and linting
