### 2022-10-19 b471d81
Added some tests and linting for info feature
