### 2022-10-19 b1cb5a2
Added test for example and fixed bug with hashing
