### 2022-10-19 d5caf61
Added test for example and fixed bug with hashing
