### 2022-10-19 b4fb785
Fixed test that was not fully implemented
