+++ tests/test_examples/test_all_examples.py
@@ -123,4 +123,4 @@ def test_composite_objects():
 def test_optimization_info():
     from examples.other_features._03_optimization_info import optimizer
 
-    assert len(optimizer.optimization_info["thresholds"]) == 50
+    assert len(optimizer.optimization_info_["all_thresholds"]) == 935
