### 2022-10-19 142ce07
Added tests for NotImplemented Error of Opti pipeline
