### 2022-10-19 a8fa56b
Fixed bug when opti pipline was called multiple times
