### 2022-10-20 cfbc19b
Experimental implementation of the clone_param method
