### 2022-10-21 bc8f997
Added some more docu for the clone param stuff
