### 2022-11-14 42f31ca
Added option to overwrite object representaiton per parameter
