+++ tpcp/_base.py
@@ -722,6 +722,9 @@ class CloneFactory(BaseFactory, Generic[T]):
         """Clone the default value for each instance."""
         return clone(self.default_value)
 
+    def __repr__(self) -> str:
+        return f"cf({self.default_value:r})"
+
 
 # The typing here is obviously wrong and a hack to make object that are wrapped by clone factory seem to be just of
 # the same type.
