+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   This can be helpful, if for some reason objects don't behave well with deepcopy.
 - Added a new method called `__repr_parameters__` that gives a class control over how params are represented.
   This can be used to customize the representation of individual parameters in the `__repr__` method.
+- Add proper repr for `CloneFactory`
 
 ## [0.11.0] - 2022-10-17
 
