### 2022-11-14 d232f2b
Fixed bug with the CloneFactory workflow
