### 2022-12-06 #4953 @xtruan
My review is complete without any major issues. The only minor suggestion I would make is setting the GitHub project's "website" field (in "About") to the docs URL (`https://tpcp.readthedocs.io/`) to make it easier to find.

This is a very useful package and after reviewing, I anticipate actually using it in my future research. Thanks!
