### 2022-12-14 30bb1b9
safe_run does not double-wrap the run method anymore
