+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [0.12.1]
+## [0.12.1] - 2022-12-14
 
 ### Changed
 - The `safe_run` method did unintentionally double-wrap the run method, if it already had a `make_action_safe` 
