+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.2] - 2022-12-14
+
+### Fixed
+- The previous for fixing hashing of objects defined in the __main__ module was not working
+  This should now be fixed.
 
 ## [0.12.1] - 2022-12-14
 
