+++ tpcp/optimize/optuna.py
@@ -1,4 +1,6 @@
 """Implementation of methods and classes to wrap the optimization Framework `Optuna`."""
+import multiprocessing
+
 try:
     import optuna  # noqa: unused-import
 except ImportError as e:
@@ -10,6 +12,7 @@ import dataclasses
 import warnings
 from typing import Any, Callable, Dict, List, Optional, Sequence, TypeVar, Union
 
+import joblib
 import numpy as np
 from optuna import Study, Trial
 from optuna.study.study import ObjectiveFuncType
@@ -24,15 +27,24 @@ from tpcp.optimize import Optimize
 
 __all__ = ["CustomOptunaOptimize", "CustomOptunaOptimizeT"]
 
-warnings.warn(
-    "The Optuna interface in tpcp is still experimental and we are testing if the workflow makes "
-    "sense even for larger projects. "
-    "This means, the interface for `CustomOptunaOptimize` will likely change in the future.",
-    UserWarning,
-)
+if multiprocessing.parent_process() is None:
+    # We want to avoid spamming the user with warnings if they are running multiple processes
+    warnings.warn(
+        "The Optuna interface in tpcp is still experimental and we are testing if the workflow makes "
+        "sense even for larger projects. "
+        "This means, the interface for `CustomOptunaOptimize` will likely change in the future.",
+        UserWarning,
+    )
 
 
 CustomOptunaOptimizeT = TypeVar("CustomOptunaOptimizeT", bound="_CustomOptunaOptimize")
+
+
+def _split_trials(n_trials, n_jobs):
+    n_per_job, remaining = divmod(n_trials, n_jobs)
+    for _ in range(n_jobs):
+        yield n_per_job + (1 if remaining > 0 else 0)
+        remaining -= 1
 
 
 class _CustomOptunaOptimize(BaseOptimize[PipelineT, DatasetT]):
@@ -47,6 +59,7 @@ class _CustomOptunaOptimize(BaseOptimize[PipelineT, DatasetT]):
     callbacks: Optional[List[Callable[[Study, FrozenTrial], None]]]
     gc_after_trial: bool
     show_progress_bar: bool
+    n_jobs: int
 
     optimized_pipeline_: PipelineT
     study_: Study
@@ -122,14 +135,61 @@ class _CustomOptunaOptimize(BaseOptimize[PipelineT, DatasetT]):
         self.dataset = dataset
 
         objective = self._create_objective(self.pipeline, dataset=dataset)
-        study = self.create_study()
-        self.study_ = self._call_optimize(study, objective)
+        self.study_ = self.create_study()
+
+        if not isinstance(self.study_._storage, optuna.storages.InMemoryStorage):
+            warnings.warn(
+                "You are using a persistent storage for the study. "
+                "This means, that the study will be saved and likely be available after this script "
+                "terminates. "
+                "To avoid issues and make sure that runs of your optimizations are reproducible and "
+                "independent, make sure to cleanup the persistent storage once the results are not needed "
+                "anymore.\n"
+                "You can use use `optuna.delete_study(study_name=opti_instance.study_.study_name, "
+                "storage=opti_instance.study_._storage)`. "
+                "Note that all result object that depend on the study are not available anymore after deletion."
+            )
+
+        if self.n_jobs == 1:
+            self._call_optimize(self.study_, objective)
+        else:
+            if isinstance(self.study_._storage, optuna.storages.InMemoryStorage):
+                raise ValueError(
+                    "You are using the InMemoryStorage with n_jobs > 1. "
+                    "This will lead to problems as the storage is not persistent and thread safe. "
+                    "Use a persistent database based storage instead."
+                )
+            if self.timeout is not None:
+                raise ValueError(
+                    "You are using timeout with n_jobs > 1. "
+                    "This parameter does not make sense in this case. "
[…]
