### 2022-12-19 d88c875
Fixed optuna load study call in case of multiprocessing. All processes now correctly use the same sampleer and pruner
