+++ tests/test_pipelines/test_optuna_optimize.py
@@ -1,3 +1,4 @@
+import time
 from typing import Callable, List, Optional, Sequence, Union
 from unittest.mock import Mock, patch
 
