### 2023-01-06 mad-lab-fau/tpcp#51 @paxtonfitzpatrick
broken links in docs
re: openjournals/joss-reviews#4953

Hey @AKuederle, thanks for creating such a useful package and documenting everything so extensively! I found the "User Guides" and "Examples" sections particularly helpful. I came across just a few minor issues with the docs. The following hyperlinked text leads to readthedocs 404 pages:
- On the [General Concepts](https://tpcp.readthedocs.io/en/latest/guides/general_concepts.html) page:
  - "[mutable handling](https://tpcp.readthedocs.io/en/latest/guides/mutable_defaults)" (second paragraph)
  - "[optimization guide](https://tpcp.readthedocs.io/en/latest/guides/optimization)" (end of the "Parameters" section)
  - "[our example](https://tpcp.readthedocs.io/en/latest/guides/compositeobjects)" (end of the "Composite Parameters" section)
- On the [Optimization and Training](https://tpcp.readthedocs.io/en/latest/guides/optimization.html) page:
  - "[evaluation guide](https://tpcp.readthedocs.io/en/latest/guides/algorithm_evaluation)" (<span>#</span>3 under "Parameter Annotations")
  - "[example](https://tpcp.readthedocs.io/en/latest/guides/optimize_pipelines)" (before the **Note** in the "Parameter Annotations" section)
