### 2023-01-06 mad-lab-fau/tpcp#52 @paxtonfitzpatrick
TODO notes left in docs
re: openjournals/joss-reviews#4953

There are a handful of leftover `TODO` notes that appear in the rendered documentation. Some of these might be reasonable to leave in for now if they refer to not-yet-implemented features or examples (e.g., the 4th & 5th ones below?), but I think the first 3 are probably worth resolving in order to make the documentation complete.

- [ ] "TODO: add cites" under the "**Group 1**" subheading in the "[Categories of Algorithms](https://tpcp.readthedocs.io/en/latest/guides/algorithm_evaluation.html#categories-of-algorithms)" section of the [Algorithm Evaluation and Parameter Optimization](https://tpcp.readthedocs.io/en/latest/guides/algorithm_evaluation.html) page
- [ ] "TODO: Link" in the **Warning** in the docstring for [`tpcp.clone`](https://tpcp.readthedocs.io/en/latest/modules/generated/parameter/tpcp.clone.html#tpcp.clone)
- [ ] "learn more TODO" in the docstring for [`tpcp.PureParameter`](https://tpcp.readthedocs.io/en/latest/modules/generated/parameter/tpcp.PureParameter.html#tpcp.PureParameter)/[`tpcp.PurePara`](https://tpcp.readthedocs.io/en/latest/modules/generated/parameter/tpcp.PurePara.html#tpcp.PurePara)
  - (also, "`parameter`" and "`pure_paramter`" appear here in snake case&mdash;should these be camel case (i.e., "`Parameter`" and "`PureParamter`" to refer to the annotation classes?)
- [ ] "Todo: Full dedicated example for `PureParameter`" under the "[The Pipeline](https://tpcp.readthedocs.io/en/latest/auto_examples/parameter_optimization/_03_gridsearch_cv.html#the-pipeline)" heading in the [`GridSearchCV` example](https://tpcp.readthedocs.io/en/latest/auto_examples/parameter_optimization/_03_gridsearch_cv.html)
- [ ] "TODO: Make GridSearchCV aware of it!" near the end of the ["Optimization Info" example](https://tpcp.readthedocs.io/en/latest/auto_examples/other_features/_03_optimization_info.html)
