### 2023-01-06 mad-lab-fau/tpcp#53 @paxtonfitzpatrick
CI tests for Python 3.11
re: openjournals/joss-reviews#4953

Based on the Python version requirement in `pyproject.toml` https://github.com/mad-lab-fau/tpcp/blob/d88c875bd33f8c0b0ea9eb6bc8b55cd8802311ed/pyproject.toml#L16 and the PyPI classifiers that Poetry automatically generates for the [packages listing](https://pypi.org/project/tpcp/) based on this, `tpcp` appears to support Python 3.8, 3.9, 3.10, and 3.11. However, the CI tests are currently run only on Python 3.8, 3.9, and 3.10: https://github.com/mad-lab-fau/tpcp/blob/d88c875bd33f8c0b0ea9eb6bc8b55cd8802311ed/.github/workflows/test-and-lint.yml#L14-L15

I'm guessing may be because Pytorch doesn't yet support Python 3.11 (pytorch/pytorch#86566) and is used in some of the tests. However, since the core of `tpcp` itself _seems_ to support Python 3.11 just fine (I've now run most of the examples with Python 3.11.0 without issue, though of course `pip install 'tpcp[torch]'` fails), Pytorch is an optional dependency, and the other optional dependencies (`optuna` & `attrs`) *do* support Python 3.11, it would be great to have CI tests run against that version if it's going to be supported.

You could do this by:
1. adding `"3.11"` to your `python-version` array in line 15 of your workflow file
2. editing your "Install dependencies" step from: https://github.com/mad-lab-fau/tpcp/blob/d88c875bd33f8c0b0ea9eb6bc8b55cd8802311ed/.github/workflows/test-and-lint.yml#L23-L28 to:
   ```yaml
    - name: Install dependencies
      env:
        PYTHON_VERSION: ${{ matrix.python-version }}
      run: |
        python -m pip install --upgrade pip
        pip install poetry
        poetry config virtualenvs.in-project true
        poetry install -E optuna
        [[ "$PYTHON_VERSION" == "3.11" ]] || poetry install -E torch
   ```
3. editing line 3 of `test_hash.py` from: https://github.com/mad-lab-fau/tpcp/blob/84a3048ff54b5d6912068009d15c65fdc5eb4592/tests/test_hash.py#L3 to:
   ```python
   torch = pytest.importorskip("torch")
   ```

Once Pytorch adds support for Python 3.11, you could simply revert steps 2 and 3.

I also think it would be worth adding a quick note to the [**Dev Setup** section](https://github.com/mad-lab-fau/tpcp/blob/main/README.md#dev-setup) of the README saying that development requires python>=3.8,<3.11 since the instructions there include installing Pytorch.

If the CI tests end up revealing other, more extensive issues with Python 3.11 compatibility that I didn't run into, you could of course instead just reduce the maximum supported version in your `pyproject.toml` for the JOSS submission.
