### 2023-01-06 d1311ca
Fixed warning during doc build
