+++ docs/conf.py
@@ -110,7 +110,7 @@ autosummary_generate_overwrite = True
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "templates"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "templates", "diagrams/README.md"]
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.
