### 2023-01-06 fd4e7f8
Resolved visible todos or turned them into github issues
