+++ docs/guides/general_concepts.rst
@@ -1,5 +1,6 @@
 General Concepts
 ================
+.. _general_concepts:
 
 Parameters
 ----------
