### 2023-01-06 cc66259
Split Hasher tests, so that only the torch tests can be skipped
