+++ .github/workflows/test-and-lint.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.8, 3.9, "3.10"]
+        python-version: [3.8, 3.9, "3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v2
@@ -25,7 +25,8 @@ jobs:
         python -m pip install --upgrade pip
         pip install poetry
         poetry config virtualenvs.in-project true
-        poetry install -E torch -E optuna
+        poetry install -E optuna
+        [[ "$PYTHON_VERSION" != "3.11" ]] || poetry install -E torch
     - name: Testing
       run: |
         poetry run poe test
