+++ .github/workflows/test-and-lint.yml
@@ -26,7 +26,7 @@ jobs:
         pip install poetry
         poetry config virtualenvs.in-project true
         poetry install -E optuna
-        [[ "$PYTHON_VERSION" != "3.11" ]] || poetry install -E torch
+        [[ "$PYTHON_VERSION" == "3.11" ]] || poetry install -E torch
     - name: Testing
       run: |
         poetry run poe test
