+++ .github/workflows/test-and-lint.yml
@@ -25,8 +25,8 @@ jobs:
         python -m pip install --upgrade pip
         pip install poetry
         poetry config virtualenvs.in-project true
-        poetry install -E optuna
-        [[ "$PYTHON_VERSION" == "3.11" ]] || poetry install -E torch
+        [[ "$PYTHON_VERSION" != "3.11" ]] || poetry install -E optuna
+        [[ "$PYTHON_VERSION" == "3.11" ]] || poetry install -E torch -E optuna
     - name: Testing
       run: |
         poetry run poe test
