+++ .github/workflows/test-and-lint.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.8, 3.9, "3.10", "3.11"]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v2
@@ -20,13 +20,20 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
+    - name: Env Setup
       run: |
         python -m pip install --upgrade pip
         pip install poetry
         poetry config virtualenvs.in-project true
-        [[ "$PYTHON_VERSION" != "3.11" ]] || poetry install -E optuna
-        [[ "$PYTHON_VERSION" == "3.11" ]] || poetry install -E torch -E optuna
+    - name: Install dependencies (<3.11)
+      if: ${{ matrix.python-version != '3.11' }}
+      run: |
+          poetry install -E optuna -E torch
+    - name: Install dependencies (3.11)
+      # In this case we need to skip the installation of torch
+      if: ${{ matrix.python-version == '3.11' }}
+      run: |
+          poetry install -E optuna
     - name: Testing
       run: |
         poetry run poe test
