### 2023-01-06 #4953 @paxtonfitzpatrick
Hey @AKuederle, thanks for putting together such a great package! I think it has the potential to be extremely useful and I'm definitely looking forward to using it in my own work. The user guides in the documentation were excellent, and I also thought the `CloneFactory` class/`cf` wrapper was a clever idea. Thanks as well for adding the readthedocs link to the repo homepage like @xtruan suggested&mdash;that's definitely a helpful improvement. I've opened a few issues above with relatively minor requests/suggestions for the docs and CI tests, but the main code itself all looks great to me.

There were two other extremely small things I noticed that I didn't feel warranted opening issues for or including in my review, but thought I'd just quickly mention here in case they're helpful:
- On line 54 of the paper (line 79 of the markdown file), I believe `[@Barth2013]` should be `@Barth2013`
- In the "Contributing" section of the README, I noticed that "opening a new issue" links to https://github.com/mad-lab-fau/tpcp/issues/new/choose, which would allow users to choose an [issue template](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/about-issue-and-pull-request-templates) to use for a new issue rather than start with a blank one, but there aren't any templates in your `.github` folder. The link obviously works perfectly fine as-is and to be clear, this is *not* a change I'm requesting you make as part of my review&mdash;just wanted to make sure you hadn't intended there to be templates available but they were accidentally deleted somehow!

Thanks again, and I'm signed off pending issues 51-53!
