### 2023-01-06 1ee69c1
UPdated ignore messages, as things seem to have changed with new prospector version
