+++ tpcp/_dataset.py
@@ -149,7 +149,10 @@ class _Dataset(BaseTpcpObject):
         """
         if (
             list(
-                map(lambda x: x is None or (isinstance(x, dict) and len(x) == 0), (groups, index, bool_map, kwargs),)
+                map(
+                    lambda x: x is None or (isinstance(x, dict) and len(x) == 0),
+                    (groups, index, bool_map, kwargs),
+                )
             ).count(False)
             > 1
         ):
@@ -531,7 +534,10 @@ class Dataset(_Dataset):
     """
 
     def __init__(
-        self, *, groupby_cols: Optional[Union[List[str], str]] = None, subset_index: Optional[pd.DataFrame] = None,
+        self,
+        *,
+        groupby_cols: Optional[Union[List[str], str]] = None,
+        subset_index: Optional[pd.DataFrame] = None,
     ):
         self.groupby_cols = groupby_cols
         self.subset_index = subset_index
