+++ tpcp/validate/_scorer.py
@@ -255,7 +255,11 @@ class Scorer(Generic[PipelineT, DatasetT, T]):
             scores.append(score)
             if self._single_score_callback:
                 self._single_score_callback(
-                    step=i, scores=tuple(scores), scorer=self, pipeline=pipeline, dataset=dataset,
+                    step=i,
+                    scores=tuple(scores),
+                    scorer=self,
+                    pipeline=pipeline,
+                    dataset=dataset,
                 )
             datapoints.append(d)
 
