### 2023-01-06 52a2145
Fixed some more linting issues and pslit workflow to only lint at one Python version
