### 2023-01-06 fad4328
Typos and better dev setup
