### 2023-01-06 70d04d9
Simplified CI
