+++ .github/workflows/test-and-lint.yml
@@ -7,7 +7,7 @@ on:
     branches: [ main ]
 
 jobs:
-  test:
+  test_and_lint:
     runs-on: ubuntu-latest
     strategy:
       matrix:
@@ -36,28 +36,11 @@ jobs:
     - name: Testing
       run: |
         poetry run poe test
+    - name: Linting
+      if: ${{ matrix.python-version == '3.9' }}
+      run: |
+        poetry run poe check
     - name: "Upload coverage to Codecov"
       uses: codecov/codecov-action@v2
       with:
         fail_ci_if_error: true
-
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python 3.9
-      uses: actions/setup-python@v2
-      with:
-        python-version: "3.9"
-    - name: Env Setup
-      run: |
-        python -m pip install --upgrade pip
-        pip install poetry
-        poetry config virtualenvs.in-project true
-    - name: Install dependencies
-      run: |
-        poetry install -E optuna -E torch
-    - name: Linting
-      run: |
-        poetry run poe check
-
