### 2023-01-06 487b04d
Finally fixed all linting issues
