+++ .prospector.yml
@@ -58,3 +58,4 @@ pylint:
     - bad-whitespace
     - cyclic-import # ignored because of false positive due to Type hints
     - import-error # We check if our dependencies are installed in the tests, no need for the linter
+    - django-not-available # Don't use it... don't know why it is checked in the first place
