### 2023-01-07 mad-lab-fau/tpcp#52 @AKuederle
Thanks again for your review!
I fully agree, many of these TODOs should be resolved.

I either resolved/implemented the changes (fd4e7f884d57d311a4714330b86a36c928b2040f) or turned the bigger once, that I might want to fix in the future into actual github issues and removed them as visibel todos form the code/docs (#56 #55)

I will close the issue. Feel free to reopen, in case you think further works should be done here.
