### 2023-01-07 mad-lab-fau/tpcp#53 @AKuederle
Thank you for the detailed instructions! I updated the CI to run tests for 3.11 as suggested

Thanks for the tip on how to skip tests based on imports. That was new to me and very helpful :)

I also updated the dev section in the README as suggested.

I will close the issue. Feel free to reopen, in case you think further work is required here.
