### 2023-01-07 #4953 @AKuederle
@paxtonfitzpatrick Thank you for that amazing review! I addressed the issues you raised (see comments in issues).
I also fixed the two minor concerns raised here (and no, there are no templates I accidentally deleted, but thanks for the concern :) )

Please check if the changes I made are sufficient in your eyes :)
