### 2023-01-07 mad-lab-fau/tpcp#52 @paxtonfitzpatrick
Thanks @AKuederle!

It looks like the third item in the list above ("learn more TODO", `parameter` vs `Parameter`, `pure_parameter` vs `PureParameter`) are still left in the docs. If you prefer to leave that page as-is for the moment because it's better/easier for you while developing, that's perfectly fine with me! I just want to make sure that wasn't accidental oversight, because I had assumed that TODO note was probably a placeholder for a link to https://tpcp.readthedocs.io/en/latest/guides/algorithm_evaluation.html#group-3-the-hybrid once that section of the docs had been finished.

Also, the 4th "Todo" in the list above (the one moved to #55) still appears in the updated version of the docs. Again, it's a tiny enough thing that I'm happy if you'd prefer to leave it in because it makes development easier&mdash;just want to make sure it was intentional because your comment seemed to indicate you meant to remove it.

Thanks again!
