### 2023-01-07 #4953 @paxtonfitzpatrick
@AKuederle thanks for addressing these so quickly! I left a follow-up comment on mad-lab-fau/tpcp#52, but your updates all look good to me! I checked off the last couple related items on my list, so my review is now complete.

Looking forward to using `tpcp` and seeing how the package grows in the future!
