### 2023-01-07 mad-lab-fau/tpcp#52 @AKuederle
I honestly missed that one :) But I will leave it in for now and will but writing a proper example/guide for the PureParameter concept on my TODO list.
