### 2023-01-07 8100a83
Updated TODO and made small correction in docs
