### 2023-01-10 1ad2a82
Added a oppinionated Optuna GridSearch equivalent
