### 2023-01-10 b4b2455
Added a oppinionated Optuna GridSearch equivalent
