+++ examples/parameter_optimization/_05_optuna_search.py
@@ -1,0 +1,137 @@
+r"""
+.. _build_in_optuna_optimizer:
+
+Build-in Optuna Optimizers
+==========================
+The :ref:`custom optuna example <custom_optuna_optimizer>` shows how to implement a specific optuna optimizer with
+full control over all aspects.
+This is still the recommended way to do things, as you often will have specific requirements for your objective
+function.
+
+However, there are still a number of problems that can be solved by a relative generic GridSearch or GridSearchCV.
+Therefore, we provide Optuna equivalents for these usecases to make use of the advanced samplers optuna provides.
+
+.. note:: We still recommend to read through the :ref:`custom optuna example <custom_optuna_optimizer>` before using
+          the specific implementations demonstrated here.
+"""
+
+# %%
+# OptunaSearch - GridSearch on Steroids
+# +++++++++++++++++++++++++++++++++++++
+# The `OptunaSearch` class can be used in all cases where you would use :class:`~tpcp.optimize.GridSearch`.
+# The following is equivalent to the GridSearch example (:ref:`grid_search`).
+
+from pathlib import Path
+
+import pandas as pd
+
+from examples.algorithms.algorithms_qrs_detection_final import QRSDetector
+from examples.datasets.datasets_final_ecg import ECGExampleData
+from tpcp import Parameter, Pipeline, cf
+
+try:
+    HERE = Path(__file__).parent
+except NameError:
+    HERE = Path(".").resolve()
+data_path = HERE.parent.parent / "example_data/ecg_mit_bih_arrhythmia/data"
+example_data = ECGExampleData(data_path)
+
+
+class MyPipeline(Pipeline[ECGExampleData]):
+    algorithm: Parameter[QRSDetector]
+
+    r_peak_positions_: pd.Series
+
+    def __init__(self, algorithm: QRSDetector = cf(QRSDetector())):
+        self.algorithm = algorithm
+
+    def run(self, datapoint: ECGExampleData):
+        # Note: We need to clone the algorithm instance, to make sure we don't leak any data between runs.
+        algo = self.algorithm.clone()
+        algo.detect(datapoint.data["ecg"], datapoint.sampling_rate_hz)
+
+        self.r_peak_positions_ = algo.r_peak_positions_
+        return self
+
+
+pipe = MyPipeline()
+
+# %%
+# Optuna Study
+# ------------
+# To use optuna we need to create an optuna study, or rather a function that returns one, that can be used by
+# `OptunaSearch` to create it.
+# We will set this up identical to the :ref:`custom optuna example <custom_optuna_optimizer>`.
+#
+# .. note:: We use a in-memory study here, if you want to use multiprocessing or ensure that your search can be
+#           continued, use a different study backend.
+from optuna import Trial, create_study, samplers
+
+
+def get_study():
+    # We use a simple RandomSampler, but every optuna sampler will work
+    sampler = samplers.RandomSampler(seed=42)
+    return create_study(direction="maximize", sampler=sampler)
+
+
+# %%
+# Search Space
+# ------------
+# In contrast to `GridSearch` where we define a fix parameter grid, in optuna we define a search space.
+# Which value sin this search space will actually be evaluated depends on the chosen sampler.
+# This also needs to be a function that takes the current trial object as input.
+def create_search_space(trial: Trial):
+    trial.suggest_float("algorithm__min_r_peak_height_over_baseline", 0.1, 2, step=0.1)
+    trial.suggest_float("algorithm__high_pass_filter_cutoff_hz", 0.1, 2, step=0.1)
+
+
+# %%
+# Score
+# -----
+# We use the same scoring function as in the `GridSearch` example:
[…]
