+++ tpcp/optimize/optuna.py
@@ -1,5 +1,4 @@
 """Implementation of methods and classes to wrap the optimization Framework `Optuna`."""
-import multiprocessing
 
 try:
     import optuna
@@ -8,14 +7,16 @@ except ImportError as e:
         "To use the tpcp Optuna interface, you first need to install optuna (`pip install optuna`)"
     ) from e
 
+import multiprocessing
 import warnings
+from collections import defaultdict
 from typing import Any, Callable, Dict, List, Optional, Sequence, TypeVar, Union
 
 import joblib
 import numpy as np
-from optuna import Study, Trial
 from optuna.study.study import ObjectiveFuncType
-from optuna.trial import FrozenTrial
+from optuna.trial import FrozenTrial, Trial
+from optuna import Study
 from typing_extensions import Self
 
 from tpcp import OptimizablePipeline, clone
@@ -23,8 +24,9 @@ from tpcp._dataset import DatasetT
 from tpcp._optimize import BaseOptimize
 from tpcp._pipeline import PipelineT
 from tpcp.optimize import Optimize
+from tpcp.validate._scorer import ScorerTypes, _validate_scorer
 
-__all__ = ["CustomOptunaOptimize", "CustomOptunaOptimizeT"]
+__all__ = ["CustomOptunaOptimize", "CustomOptunaOptimizeT", "OptunaSearch"]
 
 if multiprocessing.parent_process() is None:
     # We want to avoid spamming the user with warnings if they are running multiple processes
@@ -344,12 +346,11 @@ class CustomOptunaOptimize(_CustomOptunaOptimize[PipelineT, DatasetT]):
         params
             A dictionary representing all parameters.
         state
-            Whether the trial was completed, pruned, or any type of error occured
-        user_attrs
+            Whether the trial was completed, pruned, or any type of error occurred
         datetime_start
             When the trial was started
         datetime_complete
-            When the trial endend
+            When the trial ended
         duration
             The duration of the trial
         user_attrs_...
@@ -417,8 +418,7 @@ class CustomOptunaOptimize(_CustomOptunaOptimize[PipelineT, DatasetT]):
     optuna.
 
     Further, the use of `show_progress_bar` is not recommended when using multiprocessing, as one progress bar per
-    process is
-    created and the output is not very readable.
+    process is created and the output is not very readable.
     It might still be helpful to see that something is happening.
 
     .. note:: Using SQLite as backend is known to cause issues with multiprocessing, when the database is
@@ -508,3 +508,259 @@ class CustomOptunaOptimize(_CustomOptunaOptimize[PipelineT, DatasetT]):
             study_: Study = field(init=False, repr=False)
 
         return CustomOptunaOptimizeAt
+
+
+T = TypeVar("T")
+
+
+class OptunaSearch(CustomOptunaOptimize[PipelineT, DatasetT]):
+    """GridSearch equivalent using Optuna.
+
+    An opinionated parameter optimization for simple (i.e. non-optimizable) pipelines that can be used as a
+    replacement to GridSearch.
+
+    Parameters
+    ----------
+    pipeline
+        A tpcp pipeline with some hyper-parameters that should be optimized.
+        This should be a normal (i.e. non-optimizable pipeline) when using this class.
+    create_study
+        A callable that returns an optuna study instance to be used for the optimization.
+        It will be called as part of the `optimize` method without parameters.
+        The resulting study object can be accessed via `self.study_` after the optimization is finished.
+        Creating the study is handled via a callable, instead of providing the study object itself, to make it
+        possible to create individual studies, when CustomOptuna optimize is called by an external wrapper
+        (i.e. `cross_validate`).
+    create_search_space
+        A callable that takes a :class:`~optuna.trial.Trial` object as input and calls `suggest_*` methods on it to
[…]
