### 2023-01-10 4875ee9
Changed the baseclass structure for the new optuna method
