### 2023-01-10 b7291e9
Added test for new example
