### 2023-01-10 763af86
Made extraction of custom paras more defensive
