### 2023-01-10 3ca2b4e
Added metatest for new Optuna class
