### 2023-01-23 b370abe
It is now possible for custom aggregators to specify if the raw results should be returned
