### 2023-01-23 fd1e9c4
Gridsearch and GridserachCV both handle custom aggregator correctly now and will only try to include cols in the final output that really exist
