+++ tpcp/_utils/_general.py
@@ -42,6 +42,9 @@ def _aggregate_final_results(results: List) -> Dict:
 
 def _normalize_score_results(scores: List, prefix="", single_score_key="score"):
     """Create a scoring dictionary based on the type of `scores`."""
+    if scores[0] is None:
+        # This is the case, when we have a custom aggregator that sets RETURN_RAW_SCORES to False.
+        return None
     if isinstance(scores[0], dict):
         # multimetric scoring
         return {prefix + k: v for k, v in _aggregate_final_results(scores).items()}
