### 2023-01-23 d718a82
Added example for new Aggregator feature
