### 2023-01-23 7cfbb5f
Fixed example in optuna optimize example
