+++ .github/workflows/test-and-lint.yml
@@ -41,6 +41,7 @@ jobs:
       run: |
         poetry run poe check
     - name: "Upload coverage to Codecov"
+      if: ${{ matrix.python-version == '3.9' }}
       uses: codecov/codecov-action@v2
       with:
         fail_ci_if_error: true
