+++ docs/conf.py
@@ -38,7 +38,7 @@ project = info["name"]
 author = ", ".join(info["authors"])
 release = info["version"]
 
-copyright = "2021 - {}, MaD Lab, FAU, Digital Health and Gait Analysis Group".format(datetime.now().year)
+copyright = f"2021 - {datetime.now().year}, MaD Lab, FAU, Digital Health and Gait Analysis Group"
 
 # -- Copy the README and Changelog and fix links --------------------------------------
 
@@ -48,8 +48,8 @@ def convert_github_links(base_url, text):
 
     def substitute(matchobj):
         if matchobj.group(1) == "commit":
-            return "[{}]({})".format(matchobj.group(2)[:5], matchobj.group(0))
-        return "[#{}]({})".format(matchobj.group(2), matchobj.group(0))
+            return f"[{matchobj.group(2)[:5]}]({matchobj.group(0)})"
+        return f"[#{matchobj.group(2)}]({matchobj.group(0)})"
 
     return re.sub(regex, substitute, text)
 
@@ -154,7 +154,7 @@ intersphinx_module_mapping = {
 user_agent = "Mozilla/5.0 (X11; Linux x86_64; rv:25.0) Gecko/20100101 Firefox/25.0"
 
 intersphinx_mapping = {
-    "python": ("https://docs.python.org/{.major}".format(sys.version_info), None),
+    "python": (f"https://docs.python.org/{sys.version_info.major}", None),
     **intersphinx_module_mapping,
 }
 
