+++ tests/mixins/test_algorithm_mixin.py
@@ -11,7 +11,7 @@ from tpcp._base import BaseTpcpObjectObjT, get_param_names
 
 class TestAlgorithmMixin:
     algorithm_class = None
-    _ignored_names = tuple()
+    _ignored_names = ()
     __test__ = False
 
     @pytest.fixture()
@@ -35,7 +35,7 @@ class TestAlgorithmMixin:
     def test_all_parameters_documented(self):
         docs = NumpyDocString(inspect.getdoc(self.algorithm_class))
 
-        documented_names = set(p.name for p in docs["Parameters"])
+        documented_names = {p.name for p in docs["Parameters"]}
         actual_names = set(get_param_names(self.algorithm_class))
 
         actual_names -= set(self._ignored_names)
@@ -47,7 +47,7 @@ class TestAlgorithmMixin:
             pytest.skip("The testclass did not implement the correct `after_action_instance` fixture.")
         docs = NumpyDocString(inspect.getdoc(self.algorithm_class))
 
-        documented_names = set(p.name for p in docs["Attributes"])
+        documented_names = {p.name for p in docs["Attributes"]}
         actual_names = set(get_results(after_action_instance).keys())
 
         assert documented_names == actual_names
@@ -57,7 +57,7 @@ class TestAlgorithmMixin:
             pytest.skip("The testclass did not implement the correct `after_action_instance` fixture.")
         docs = NumpyDocString(inspect.getdoc(self.algorithm_class))
 
-        documented_names = set(p.name for p in docs["Other Parameters"])
+        documented_names = {p.name for p in docs["Other Parameters"]}
         actual_names = set(get_action_params(after_action_instance).keys())
 
         assert documented_names == actual_names
