+++ pyproject.toml
@@ -29,7 +29,8 @@ torch = ["torch"]
 optuna = ["optuna"]
 attrs = ["attrs"]
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
+pydata-sphinx-theme = "^0.12.0"
 pytest = "^6.2.1"
 pytest-cov = "^2.8.1"
 black = "^22.3.0"
@@ -37,7 +38,6 @@ poethepoet = "^0.10.0"
 pyright = "^1.1.230"
 numpydoc = "^1.4.0"
 sphinx-gallery = "^0.11.0"
-pydata-sphinx-theme = "^0.9.0"
 recommonmark = "^0.7.1"
 memory-profiler = "^0.58.0"
 matplotlib = "^3.4.3"
