+++ docs/conf.py
@@ -132,11 +132,6 @@ html_theme_options = {
     "github_url": "https://github.com/mad-lab-fau/tpcp",
     "use_edit_page_button": True,
     "show_prev_next": False,
-    # Workaround until pydata-sphinx-theme 0.13 is released (https://github.com/pydata/pydata-sphinx-theme/issues/1094)
-    "logo": {
-        "image_light": "logo/gaitmap_logo.png",
-        "image_dark": "logo/gaitmap_logo.png",
-    }
 }
 
 html_context = {
