### 2023-01-31 a3734af
Tested and fixed bug when using GroupKFold with more than one column in create_group_labels
