### 2023-01-31 220cc00
Tested and fixed bug when using GroupKFold with more than one column in create_group_labels
