### 2023-02-01 5604735
Fixed hashing of objects in main for good
