### 2023-02-01 73801c2
Fixed hashing of objects in main for good
