### 2023-02-06 457be9b
Prepared new options for return_optimized internally
