+++ tpcp/optimize/_optimize.py
@@ -6,7 +6,7 @@ from contextlib import nullcontext
 from functools import partial
 from itertools import product
 from tempfile import TemporaryDirectory
-from typing import Any, ContextManager, Dict, Generic, Iterator, List, Optional, Tuple, TypeVar, Union
+from typing import Any, ContextManager, Dict, Generic, Iterator, List, Optional, Tuple, TypeVar, Union, Literal
 
 import numpy as np
 from joblib import Memory, delayed
@@ -273,7 +273,9 @@ class GridSearch(BaseOptimize[PipelineT, DatasetT], Generic[PipelineT, DatasetT,
             If a single score is used for scoring, then the generic name "score" is used.
             Otherwise, multiple columns with the name of the respective scorer exist
         rank_score / rank_{scorer-name}
-            A sorting for each score from the highest to the lowest value
+            A sorting for each score from the highest to the lowest value.
+            If lower or higher values are better, depends on the scoring function and needs to be interpreted
+            accordingly.
         single_score / single_{scorer-name}
             The individual scores per data point for each parameter combination.
             This is a list of values with the `len(dataset)`.
@@ -375,20 +377,19 @@ class GridSearch(BaseOptimize[PipelineT, DatasetT], Generic[PipelineT, DatasetT,
         # have handled issues with non-uniform cases already.
         first_test_score = results[0]["scores"]
         self.multimetric_ = isinstance(first_test_score, dict)
-        _validate_return_optimized(self.return_optimized, self.multimetric_, first_test_score)
 
         results = self._format_results(
             list(self.parameter_grid),
             results,
         )
 
-        if self.return_optimized:
-            return_optimized = "score"
-            if self.multimetric_ and isinstance(self.return_optimized, str):
-                return_optimized = self.return_optimized
-            self.best_index_ = results[f"rank_{return_optimized}"].argmin()
-            self.best_score_ = results[return_optimized][self.best_index_]
-            self.best_params_ = results["params"][self.best_index_]
+        reverse_ranking, return_optimized = _validate_return_optimized(
+            self.return_optimized, self.multimetric_, first_test_score
+        )
+        if return_optimized:
+            self.best_index_, self.best_score_, self.best_params_ = _extract_return_optimize_info(
+                return_optimized, reverse_ranking, results, rank_prefix="rank_", score_prefix=""
+            )
             # We clone twice, in case one of the params was itself an algorithm.
             self.optimized_pipeline_ = self.pipeline.clone().set_params(**self.best_params_).clone()
 
@@ -749,14 +750,13 @@ class GridSearchCV(BaseOptimize[OptimizablePipelineT, DatasetT], Generic[Optimiz
 
         first_test_score = out[0]["test_scores"]
         self.multimetric_ = isinstance(first_test_score, dict)
-        _validate_return_optimized(self.return_optimized, self.multimetric_, first_test_score)
-        if self.return_optimized:
-            return_optimized = "score"
-            if self.multimetric_ and isinstance(self.return_optimized, str):
-                return_optimized = self.return_optimized
-            self.best_index_ = results[f"rank_test_{return_optimized}"].argmin()
-            self.best_score_ = results[f"mean_test_{return_optimized}"][self.best_index_]
-            self.best_params_ = results["params"][self.best_index_]
+        reverse_ranking, return_optimized = _validate_return_optimized(
+            self.return_optimized, self.multimetric_, first_test_score
+        )
+        if return_optimized:
+            self.best_index_, self.best_score_, self.best_params_ = _extract_return_optimize_info(
+                return_optimized, reverse_ranking, results, rank_prefix="rank_test_", score_prefix="mean_test_"
+            )
             # We clone twice, in case one of the params was itself an algorithm.
[…]
