### 2023-02-06 f38ea44
Added option to prefix return_optimize with a `-` sign to invert scoring selection
