### 2023-02-06 9d29e18
Added option to prefix return_optimize with a `-` sign to invert scoring selection
