+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.0] - 
+
+### Added
+- GridSearch and GridSearchCV now have the option to pick the parameters with the lowest score if desired.
+  This is useful, if your metric represents an error and you want to pick the parameters that minimize the error.
+  To do that, you can set the `return_optimized` parameter of these classes to the name of metric prefixed with a `-`.
+  (e.g. `return_optimized="-rmse"`).
+  (https://github.com/mad-lab-fau/tpcp/pull/61)
+
 ## [0.14.0] - 2023-02-01
 
 ## Added
