### 2023-02-06 #4953 @danielskatz
@openjournals/dev - please take a look at the warning ☝️ and advise @openjournals/dsais-eics and @adi3 if any changes are needed
