### 2023-02-06 644fa57
Added check for deeterministic index
