### 2023-02-06 ea85dfe
Added check for deeterministic index
