### 2023-02-07 6e2850f
Added further safeguards in teh crossvalidate methods
