### 2023-02-07 8762066
Added further safeguards in teh crossvalidate methods
