+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.15.0] - 
+## [0.15.0] - 2023-02-07
 
 ### Added
 - GridSearch and GridSearchCV now have the option to pick the parameters with the lowest score if desired.
