### 2023-02-13 5fbe665
WIP: Callback for optuna gridsearch
