### 2023-02-13 #4953 @arfon
> The paper's PDF and metadata files generation produced some warnings that could prevent the final paper from being published. Please fix them before the end of the review process.

These look to be errors associated with the abstracts in the paper's bibtex file. As such, I don't think they will affect the JOSS submission.
